refactor(ui): コンポーネント分割
