+++ src/view/pages/index/filter/filter-by-parts/event.ts
@@ -1,0 +1,5 @@
+import type { ReadonlyPartsFilterState } from '~core/assembly/filter/filter-set.ts'
+
+export type ChangeFilter = {
+  target: ReadonlyPartsFilterState
+}
